requirements.txt
anthropic==0.2.10
numpy==1.22.3
openai==0.27.8
pandas==1.5.3
protobuf==3.9.0
tenacity==8.2.2
tiktoken==0.4.0
google-generativeai==0.1.0